Remove channel+leagueid requirement at install time (#264)
- Installers now have to use the fplbot follow <leagueid> command

+++ src/FplBot.Tests/Helpers/InMemorySlackTeamRepository.cs
@@ -8,7 +8,7 @@ namespace FplBot.Tests.Helpers
 {
     public class InMemorySlackTeamRepository : ISlackTeamRepository
     {
-        private readonly long _leagueId;
+        private readonly int _leagueId;
 
         public InMemorySlackTeamRepository()
         {
